+++ cmd/delete.go
@@ -16,15 +16,9 @@
 package cmd
 
 import (
-	"fmt"
-	"sort"
-
-	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/ksonnet/kubecfg/utils"
+	"github.com/ksonnet/kubecfg/pkg/kubecfg"
 )
 
 const (
@@ -42,75 +36,30 @@ var deleteCmd = &cobra.Command{
 	Short: "Delete Kubernetes resources described in local config",
 	RunE: func(cmd *cobra.Command, args []string) error {
 		flags := cmd.Flags()
+		var err error
 
-		gracePeriod, err := flags.GetInt64(flagGracePeriod)
+		c := kubecfg.DeleteCmd{}
+
+		c.GracePeriod, err = flags.GetInt64(flagGracePeriod)
 		if err != nil {
 			return err
 		}
 
-		files, err := getFiles(cmd, args)
+		c.ClientPool, c.Discovery, err = restClientPool(cmd)
 		if err != nil {
 			return err
 		}
 
-		vm, err := newExpander(cmd)
+		c.DefaultNamespace, _, err = clientConfig.Namespace()
 		if err != nil {
 			return err
 		}
 
-		objs, err := vm.Expand(files)
+		objs, err := readObjs(cmd, args)
 		if err != nil {
 			return err
 		}
 
-		clientpool, disco, err := restClientPool(cmd)
-		if err != nil {
-			return err
-		}
-
-		defaultNs, _, err := clientConfig.Namespace()
-		if err != nil {
-			return err
-		}
-
-		version, err := utils.FetchVersion(disco)
-		if err != nil {
-			return err
-		}
-
-		sort.Sort(sort.Reverse(utils.DependencyOrder(objs)))
-
-		deleteOpts := metav1.DeleteOptions{}
-		if version.Compare(1, 6) < 0 {
-			// 1.5.x option
-			boolFalse := false
-			deleteOpts.OrphanDependents = &boolFalse
-		} else {
-			// 1.6.x option (NB: Background is broken)
-			fg := metav1.DeletePropagationForeground
-			deleteOpts.PropagationPolicy = &fg
-		}
-		if gracePeriod >= 0 {
-			deleteOpts.GracePeriodSeconds = &gracePeriod
-		}
-
-		for _, obj := range objs {
-			desc := fmt.Sprintf("%s %s", utils.ResourceNameFor(disco, obj), utils.FqName(obj))
-			log.Info("Deleting ", desc)
-
-			c, err := utils.ClientForResource(clientpool, disco, obj, defaultNs)
-			if err != nil {
-				return err
-			}
-
-			err = c.Delete(obj.GetName(), &deleteOpts)
-			if err != nil && !errors.IsNotFound(err) {
-				return fmt.Errorf("Error deleting %s: %s", desc, err)
-			}
-
-			log.Debug("Deleted object: ", obj)
-		}
-
-		return nil
+		return c.Run(objs)
 	},
 }
